Make from_dict a DataFrame class method
