+++ dask/dataframe/__init__.py
@@ -18,7 +18,6 @@ try:
         from_bcolz,
         from_dask_array,
         from_delayed,
-        from_dict,
         from_map,
         from_pandas,
         read_csv,
