+++ dask/dataframe/io/__init__.py
@@ -7,7 +7,6 @@ from dask.dataframe.io.io import (
     from_bcolz,
     from_dask_array,
     from_delayed,
-    from_dict,
     from_map,
     from_pandas,
     to_bag,
